Fix a compiler warning

+++ src/ConfigurableFirmata.cpp
@@ -460,7 +460,7 @@ void FirmataClass::sendSysex(byte command, byte bytec, byte *bytev)
  */
 void FirmataClass::sendString(byte command, const char *string)
 {
-  sendSysex(command, strlen(string), (byte *)string);
+  sendSysex(command, (byte)strlen(string), (byte *)string);
 }
 
 /**
